Fixed normalization

+++ include/BALL/MATHS/FFT3D.h
@@ -1,7 +1,7 @@
 // -*- Mode: C++; tab-width: 2; -*-
 // vi: set ts=2:
 //
-// $Id: FFT3D.h,v 1.12.2.2 2005/06/08 17:59:50 oliver Exp $
+// $Id: FFT3D.h,v 1.12.2.3 2006/01/03 19:08:08 anhi Exp $
 //
 
 #ifndef BALL_MATHS_TFFT3D_H
@@ -1127,7 +1127,7 @@ namespace BALL
 																		from.getPhysSpaceMaxZ()));
 
 			// and fill it
-			double normalization = 1./(pow((float)(lengthX*lengthY*lengthZ),from.getNumberOfInverseTransforms()));
+			double normalization = 1./(pow((float)(lengthX*lengthY*lengthZ),(int)from.getNumberOfInverseTransforms()));
 			typename TFFT3D<ComplexTraits>::Complex dataIn;
 			typename TFFT3D<ComplexTraits>::Complex dataOut;
 			
@@ -1177,7 +1177,7 @@ namespace BALL
 
 			// and fill it
 			// AR: old version double normalization=1./(sqrt(2.*M_PI))*(stepPhysX*stepPhysY*stepPhysZ)/(pow((float)(lengthX*lengthY*lengthZ),from.getNumberOfInverseTransforms()));
-			double normalization=1./pow(sqrt(2.*M_PI),3)/(pow((float)(lengthX*lengthY*lengthZ),from.getNumberOfInverseTransforms()));
+			double normalization=1./pow(sqrt(2.*M_PI),3)/(pow((float)(lengthX*lengthY*lengthZ),(int)from.getNumberOfInverseTransforms()));
 			
 			Index x, y, z;
 			signed int xp, yp, zp;
